feat: Checkout Kata MVP is complete
The requirements of the Checkout Kata have been satisfied.
The Project does not provide an implementation of a Pricer
but it allows for an implementation (or several
different implementations) to be created elsewhere and
plugged in using Dependency Injection.

For now, the Unit Test use mocking to satisfy the
dependency on a notional Pricer.